stores: move type-only exports out of widget-store
WidgetStep, the Rebuild* types, and the pref-key unions move to
src/types/; SettingsTab consolidates on the ui-state definition; the
ipc re-export block is dropped in favor of direct @/ipc/types imports.
Step F1a of docs/2026-06-12-viewmodel-completion-plan.md (Stage 0):
widget-store's surface shrinks so its later deletion is pure file
removal.
