+++ apps/native/src/types/preferences.ts
@@ -1,0 +1,8 @@
+export type ConfirmPrefKey = "confirmBuild" | "confirmClear" | "confirmRollback";
+
+export type BoolPrefKey =
+  | ConfirmPrefKey
+  | "autoSummarizeOnFocus"
+  | "scanHomebrewOnStartup"
+  | "defaultToDiffTab"
+  | "experimentalSpinningMascot";
